Cleanup the hofstede dataset
The original source dataset uses a different set of three character
country codes (made up maybe?) than iso3166. In order to convert them
over to the two character codes necessary for the flag lookup, I just
went ahead and created a separate file for now and updated the country
codes. While I was in there, I separated the file into two datasets --
one for "countries" and one for other regional groupings that were
either subsets or supersets of countries.

I'm not making any political judgement here in my use of the term
countries -- I recognize that many of these are not actually countries
in and of themselves, but they happen to be included in the iso3166 list
as distinct entities with three character codes. For instance, Puerto
Rico is a United States territory.